feat: adiciona criptografia em senha, jwt, secret, cria funcao verificarToken, adiciona endpoint login e faz verificacao de permissao de usuario nos endpoints

+++ backend/server.js
@@ -1,3 +1,6 @@
+//importar dotenv
+require("dotenv").config();
+
 //importar o banco de dados
 const db = require("./database");
 
@@ -7,6 +10,12 @@ const express = require("express");
 //importar o bcrypt para criptografar a senha do usuário
 const bcrypt = require("bcrypt");
 
+//importar o jsonwebtoken
+const jwt = require("jsonwebtoken");
+
+//pegar SECRET do .env
+const SECRET = process.env.JWT_SECRET;
+
 // criar aplicação
 const app = express();
 
@@ -14,6 +23,31 @@ const app = express();
 const PORT = 3000;
 
 app.use(express.json());
+
+function verificarToken(req, res, next){
+    const authHeader = req.headers["authorization"];
+
+    if(!authHeader){
+        return res.status(401).json({
+            erro: "Token não fornecido"
+        });
+    }
+
+    const token = authHeader.split(" ")[1];
+
+    jwt.verify(token, SECRET, (err, decoded) => {
+        if(err){
+            return res.status(403).json({
+                erro: "Token inválido"
+            });
+        }
+
+        req.userId = decoded.id;
+
+        next();
+    });
+
+}
 
 // criar rota
 app.get("/", (req, res) => {
@@ -26,7 +60,7 @@ app.listen(PORT, () => {
 });
 
 //lista usuários
-app.get("/users", (req, res) => {
+app.get("/users", verificarToken, (req, res) => {
     db.all(
         "SELECT * FROM users",
         [],
@@ -50,7 +84,7 @@ app.get("/users", (req, res) => {
 });
 
 //lista um usuário específico
-app.get("/users/:id", (req, res) => {
+app.get("/users/:id", verificarToken, (req, res) => {
     const id = req.params.id;
     db.get(
         "SELECT * FROM users WHERE id = ?",
@@ -120,9 +154,15 @@ app.post("/users", async (req, res) => {
 });
 
 //atualiza um usuário específico
-app.put("/users/:id", async (req, res) => {
-    let id = req.params.id;
-    id = parseInt(id);
+app.put("/users/:id", verificarToken, async (req, res) => {
+    const id = parseInt(req.params.id);
+
+    //Bloqueia se tentar editar outro usuário
+    if(req.userId != id){
+        return res.status(403).json({
+            erro: "Você não tem permissão para editar este usuário"
+        });
+    }
 
     const { nome, data_nascimento, email, senha } = req.body;
 
@@ -164,11 +204,16 @@ app.put("/users/:id", async (req, res) => {
 });
 
 //Deleta um usuário específico
-app.delete("/users/:id", (req, res) => {
+app.delete("/users/:id", verificarToken, (req, res) => {
     //pegar id da URL
-    let id = req.params.id;
-    //Converter para número
-    id = parseInt(id);
+    const id = parseInt(req.params.id);
+
+    //comparar id do usuário com o forncecido pela URL
+    if(req.userId != id){
+        return res.status(403).json({
+            erro: "Você não tem permissão para deletar este usuário"
+        });
+    }
     
     db.run(
         "DELETE FROM users WHERE id = ?",
@@ -184,4 +229,54 @@ app.delete("/users/:id", (req, res) => {
             res.json({ mensagem: `Usuário com o id ${id} foi deletado!` });
         }
     );
+});
+
+app.post("/login", async (req, res) => {
+    const { email, senha} = req.body;
+
+    db.get(
+        "SELECT * FROM users WHERE email = ?",
+        [email],
+        async(err, row) => {
+            if(err){
+                console.log(err);
+                return res.json({ erro: "Erro ao buscar usuário" });
+            }
+            if(!row){
+                return res.json({ erro: "Email ou senha inválidos" });
+            }
+
+            try{
+                //comparar senha digitada com hash
+                const senhaValida = await bcrypt.compare(
+                    senha,
+                    row.senha
+                );
+
+                if(!senhaValida){
+                    return res.json({ erro: "Email ou senha inválidos" });
+                }
+
+                const token = jwt.sign(
+                    {id: row.id}, //payload
+                    SECRET, //SECRET KEY
+                    {expiresIn: "1h"} //Opções
+                );
+
+                res.json({
+                    mensagem: "Login realizado com sucesso!",
+                    token: token
+                });
+
+
+            }catch(erro){
+                console.log(erro);
+
+                res.json({
+                    erro: "Erro ao validar senha"
+                });
+            }
+        }
+    );
+
 });
